requirements.txt
numpy
pandas
matplotlib
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence_transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv